requirements.txt
darts==0.29.0
matplotlib==3.8.4
numpy==1.26.4
optuna==3.6.1
pandas==2.2.2
pillow==10.3.0
pmdarima==2.0.4
pytorch-lightning==2.2.4
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
statsforecast==1.7.4
statsmodels==0.14.2
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.3.2
torchvision==0.18.0
tqdm==4.66.4
xgboost==2.0.3
jupyter==1.0.0
wandb